enable js.erb parsing with webpacker, rename references to tinymce.js, and properly get the asset_path for the tinymce content stylesheet
