+++ app/javascript/views/guidances/new_edit.js
@@ -1,4 +1,4 @@
-import { Tinymce } from '../../utils/tinymce';
+import { Tinymce } from '../../utils/tinymce.js.erb';
 
 $(() => {
   Tinymce.init({ selector: '#guidance-text' });
