+++ app/javascript/views/org_admin/phases/new_edit.js
@@ -1,5 +1,5 @@
 import 'bootstrap-sass/assets/javascripts/bootstrap/collapse';
-import { Tinymce } from '../../../utils/tinymce';
+import { Tinymce } from '../../../utils/tinymce.js.erb';
 import { isObject, isString } from '../../../utils/isType';
 import getConstant from '../../../constants';
 import expandCollapseAll from '../../../utils/expandCollapseAll';
