+++ app/javascript/views/org_admin/templates/edit.js
@@ -1,4 +1,4 @@
-import { Tinymce } from '../../../utils/tinymce';
+import { Tinymce } from '../../../utils/tinymce.js.erb';
 import { eachLinks } from '../../../utils/links';
 import { isObject, isString } from '../../../utils/isType';
 import { renderNotice, renderAlert } from '../../../utils/notificationHelper';
