+++ app/javascript/views/org_admin/templates/new.js
@@ -1,4 +1,4 @@
-import { Tinymce } from '../../../utils/tinymce';
+import { Tinymce } from '../../../utils/tinymce.js.erb';
 import { eachLinks } from '../../../utils/links';
 
 $(() => {
