+++ app/javascript/views/orgs/admin_edit.js
@@ -1,7 +1,7 @@
 // TODO: we need to be able to swap in the appropriate locale here
 import 'number-to-text/converters/en-us';
 import { isObject } from '../../utils/isType';
-import { Tinymce } from '../../utils/tinymce';
+import { Tinymce } from '../../utils/tinymce.js.erb';
 import { eachLinks } from '../../utils/links';
 
 $(() => {
