+++ app/javascript/views/plans/edit_details.js
@@ -1,4 +1,4 @@
-import { Tinymce } from '../../utils/tinymce';
+import { Tinymce } from '../../utils/tinymce.js.erb';
 import getConstant from '../../constants';
 import 'bootstrap-3-typeahead';
