+++ app/javascript/views/super_admin/notifications/edit.js
@@ -1,4 +1,4 @@
-import { Tinymce } from '../../../utils/tinymce';
+import { Tinymce } from '../../../utils/tinymce.js.erb';
 
 $(() => {
   Tinymce.init({ selector: '.notification-text', forced_root_block: '' });
