+++ app/javascript/views/super_admin/themes/new_edit.js
@@ -1,4 +1,4 @@
-import { Tinymce } from '../../../utils/tinymce';
+import { Tinymce } from '../../../utils/tinymce.js.erb';
 
 $(() => {
   Tinymce.init({ selector: '#theme_description' });
