+++ config/webpack/environment.js
@@ -1,4 +1,5 @@
 const { environment } = require('@rails/webpacker')
+const erb =  require('./loaders/erb')
 const webpack = require('webpack')
 
 const eslint = require('./loaders/eslint')
@@ -14,4 +15,5 @@ environment.plugins.append('Provide', new webpack.ProvidePlugin({
 environment.loaders.prepend('ESLint', eslint)
 environment.loaders.append('Babel-Loader', babelLoader)
 
+environment.loaders.append('erb', erb)
 module.exports = environment
