+++ package.json
@@ -41,6 +41,7 @@
     "js-cookie": "^2.2.0",
     "moment": "^2.22.2",
     "number-to-text": "^0.3.5",
+    "rails-erb-loader": "^5.5.2",
     "timeago.js": "4.0.0-beta.1",
     "tinymce": "^4.8.3",
     "webpack": "^3.12.0",
